ChatLearner/text_to_speech/requirements.txt
google-cloud-texttospeech==0.4.0
playsound